+++ hy/importer.py
@@ -14,6 +14,7 @@ from hy.compiler import hy_compile
 from hy.reader import read_many, HyReader
 
 
+
 @contextmanager
 def loader_module_obj(loader):
     """Use the module object associated with a loader.